Fixing Google Analytics failures on new instances
This handles the case where a table cannot be fingerprinted, for
example, Google Analytics. If an error is thrown during fingerprinting,
log that error and carry on with the sync. Previously this would have
aborted the sync entirely.

Add an assert to ensure that `:metrics` is present in a GA query. The
Google library will fail with a NPE when it's not.

[ci googleanalytics]